api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
python-dotenv==1.0.1
redis==5.2.0
crewai==0.80.0
crewai-tools==0.12.0
google-generativeai==0.8.3
huggingface-hub==0.26.0
requests==2.32.3
firebase-admin==6.5.0
slowapi==0.1.9
python-multipart==0.0.12
